+++ libraries/random/random.h
@@ -25,7 +25,7 @@ uint32_t rnd_pcg32_rand_range(uint32_t lim);
 
 // Generate a sequence of random numbers 0 <= r < lim.
 void
-rnd_pcg32_rand_range_fill(uint32_t *mem, uint32_t lim, uint32_t n);
+rnd_pcg32_rand_range_fill(uint32_t *mem, uint32_t lim, size_t n);
 
 void
 rnd_pcg32_rand_uniform_fill_float(float *mem, uint32_t n);
